requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-auth==1.30.0
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.4
